requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==8.1.2
pyspark==2.4.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
